Refactor into directory modules
