+++ src/server/mod.rs
@@ -1,0 +1,7 @@
+pub mod endpoint;
+
+use super::errors;
+use super::eth;
+use super::relay;
+use super::transfers;
+use super::utils;
