+++ src/transfers/live.rs
@@ -1,6 +1,4 @@
-use std::fmt;
 use std::rc::Rc;
-
 use std::time;
 use tokio_core::reactor;
 use web3::confirm::{wait_for_transaction_confirmation, SendTransactionWithConfirmation};
@@ -10,23 +8,10 @@ use web3::futures::try_ready;
 use web3::types::{Address, FilterBuilder, TransactionReceipt, H256, U256};
 use web3::DuplexTransport;
 
-use super::contracts::TRANSFER_EVENT_SIGNATURE;
+use super::eth::contracts::TRANSFER_EVENT_SIGNATURE;
 use super::relay::{Network, TransactionApprovalState};
-use super::transaction::SendTransaction;
-use super::withdrawal::{ApproveParams, DoesRequireApproval, UnapproveParams};
-use utils::{CheckLogRemoved, CheckRemoved};
-use web3::contract::tokens::Tokenize;
-
-/// Add CheckRemoved trait to SendTransaction, which is called by Transfer::approve_withdrawal
-impl<T, P> CheckRemoved<T, (), ()> for SendTransaction<T, P>
-where
-    T: DuplexTransport + 'static,
-    P: Tokenize + Clone + 'static,
-{
-    fn check_log_removed(self, target: &Rc<Network<T>>, tx_hash: H256) -> CheckLogRemoved<T, (), ()> {
-        CheckLogRemoved::new(target.clone(), tx_hash, Box::new(self))
-    }
-}
+use super::transfers::transfer::Transfer;
+use super::utils::{CheckLogRemoved, CheckRemoved};
 
 /// Add CheckRemoved trait to SendTransactionWithConfirmation, which is returned by wait_for_transaction_confirmation()
 impl<T> CheckRemoved<T, TransactionReceipt, web3::Error> for SendTransactionWithConfirmation<T>
@@ -42,129 +27,15 @@ where
     }
 }
 
-/// Represents a token transfer between two networks
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct Transfer {
-    pub destination: Address,
-    pub amount: U256,
-    pub tx_hash: H256,
-    pub block_hash: H256,
-    pub block_number: U256,
-    pub removed: bool,
-}
-
-impl Transfer {
-    /// Returns a Transfer object based on a receipt, and some values
-    ///
-    /// # Arguments
-    ///
-    /// * `address` Address of the wallet that sent the transfer
-    /// * `amount` - Amount of ERC20 token sent
-    /// * `removed` - Was this log removed
-    /// * `receipt` - TransactionReceipt for this log
-    ///
-    pub fn from_receipt(
-        destination: Address,
-        amount: U256,
-        removed: bool,
-        receipt: &TransactionReceipt,
-    ) -> Result<Self, String> {
-        if receipt.block_number.is_none() {
-            return Err("no block number in transfer receipt".to_string());
-        }
-
-        if receipt.block_hash.is_none() {
-            return Err("no block hash in transfer receipt".to_string());
-        }
-
-        let block_hash = receipt.block_hash.unwrap();
-        let block_number = receipt.block_number.unwrap();
-
-        Ok(Transfer {
-            destination,
-            amount,
-            tx_hash: receipt.transaction_hash,
-            block_hash,
-            block_number,
-            removed,
-        })
-    }
-    /// Returns a Future that will get the Withdrawal from the contract
-    ///
-    /// # Arguments
-    ///
-    /// * `target` - Network that withdrawals are posted to
-    pub fn check_withdrawal<T: DuplexTransport + 'static>(&self, target: &Rc<Network<T>>) -> DoesRequireApproval<T> {
-        DoesRequireApproval::new(target, self)
-    }
-
-    /// Returns a Future that will transaction with "approve_withdrawal" on the ERC20Relay contract
-    ///
-    /// # Arguments
-    ///
-    /// * `target` - Network where the withdrawals is performed
-    pub fn approve_withdrawal<T: DuplexTransport + 'static>(
-        &self,
-        target: &Rc<Network<T>>,
-    ) -> Box<Future<Item = (), Error = ()>> {
-        info!("approving withdrawal on {:?}: {} ", target.network_type, self);
-        let target = target.clone();
-        Box::new(
-            SendTransaction::new(
-                &target,
-                "approveWithdrawal",
-                &ApproveParams::from(*self),
-                target.retries,
-            )
-            .check_log_removed(&target, self.tx_hash)
-            .and_then(move |success| {
-                success.map_or_else(
-                    || {
-                        warn!(
-                            "log removed from originating chain while waiting on approval confirmations on target {:?}",
-                            target.network_type
-                        );
-                        Ok(())
-                    },
-                    |_| Ok(()),
-                )
-            }),
-        )
-    }
-
-    pub fn unapprove_withdrawal<T: DuplexTransport + 'static>(
-        &self,
-        target: &Rc<Network<T>>,
-    ) -> SendTransaction<T, UnapproveParams> {
-        info!("unapproving withdrawal on {:?}: {} ", target.network_type, self);
-        SendTransaction::new(
-            target,
-            "unapproveWithdrawal",
-            &UnapproveParams::from(*self),
-            target.retries,
-        )
-    }
-}
-
-impl fmt::Display for Transfer {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "({} -> {:?}, hash: {:?})",
-            self.amount, self.destination, self.tx_hash
-        )
-    }
-}
-
 /// Stream of transfer events that have been on the main chain for N blocks.
 /// N is confirmations per settings.
-pub struct WatchTransferLogs<T: DuplexTransport + 'static> {
+pub struct WatchLiveTransfers<T: DuplexTransport + 'static> {
     handle: reactor::Handle,
     rx: mpsc::UnboundedReceiver<Transfer>,
     target: Rc<Network<T>>,
 }
 
-impl<T: DuplexTransport + 'static> WatchTransferLogs<T> {
+impl<T: DuplexTransport + 'static> WatchLiveTransfers<T> {
     /// Returns a newly created WatchTransfers Stream
     ///
     /// # Arguments
@@ -311,7 +182,7 @@ impl<T: DuplexTransport + 'static> WatchTransferLogs<T> {
 
         handle.spawn(future);
 
-        WatchTransferLogs {
+        WatchLiveTransfers {
             handle: handle.clone(),
             rx,
             target: target.clone(),
@@ -319,7 +190,7 @@ impl<T: DuplexTransport + 'static> WatchTransferLogs<T> {
     }
 }
 
-impl<T: DuplexTransport + 'static> Future for WatchTransferLogs<T> {
+impl<T: DuplexTransport + 'static> Future for WatchLiveTransfers<T> {
     type Item = ();
     type Error = ();
     fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
